requirements.txt
-e git://github.com/kingsaffair/django-ucamwebauth.git#egg=django-ucamwebauth
django
django-webpack-loader
djangorestframework
